% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_annee_spi.R
\name{lm_annee_spi}
\alias{lm_annee_spi}
\title{Modéliser des variables en fonction de l'année et de la présence du spirlin}
\usage{
lm_annee_spi(model_data, metriques, interaction = TRUE)
}
\arguments{
\item{model_data}{Dataframe contenant les données. Il contient une colonne par variable à modéliser,
une colonne annee et une colonne spi.}

\item{metriques}{Vecteur caractères. Nom des colonnes contenant les variables à modéliser.}

\item{interaction}{Booléen. L'interaction annee x spi doit-elle être incluse en variable
explicative ? Par défaut interaction = TRUE.}
}
\value{
Un dataframe résumant les modèles.
}
\description{
Permet de caler une régression multiple de chaque variable spécifiée en fonction de l'année
(variable annee) et de la présence du spirlin (variable spi codée 0/1). Les sorties des modèles sont
synthétisées et mises en forme dans un tableau.
}
\examples{
\dontrun{
# avec terme d'interaction
resultat <- lm_annee_spi(mod_data, metriques = mes_metriques)

# sans terme d'interaction
lm_annee_spi(mod_data, metriques = mes_metriques, interaction = FALSE)
}

}
