% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_and_filter_species.R
\name{recode_and_filter_species}
\alias{recode_and_filter_species}
\title{Recodage des codes espèces et exclusion de taxons}
\usage{
recode_and_filter_species(df, sp_to_remove = NA)
}
\arguments{
\item{df}{Dataframe standardisé contenant les données.}

\item{sp_to_remove}{Vecteur texte contenant les codes à trois lettres des taxons à
supprimer.}
}
\value{
Le dataframe mis à jour.
}
\description{
Par exemple les carpes cuir, miroir, etc. sont regroupées sous un unique code CCO.
}
